third_party/WebKit/LayoutTests/external/wpt/svg/shapes/animatedPoints-non-animated-expected.txt

This is a testharness.js-based test.
FAIL animatedPoints represents non-animated value assert_equals: expected 20 but got 170
Harness: the test ran to completion.
